+++ Tests/CoreSubspec/CoreSubspecSpec.swift
@@ -14,16 +14,6 @@ final class CoreSubspecSpec: QuickSpec {
             it("should not contain RAnalyticsRATTracker") {
                 expect(NSClassFromString("RAnalyticsRATTracker")).to(beNil())
             }
-
-            it("should contain SDKTracker") {
-                let trackers = AnalyticsManager.shared().trackers
-
-                expect(trackers).toNot(beNil())
-                expect(trackers.count).to(equal(1))
-
-                let sdkTrackerIsContained = trackers.first(where: { $0 is SDKTracker }) != nil
-                expect(sdkTrackerIsContained).to(beTrue())
-            }
         }
     }
 }